requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0  # For SQLite async support
redis==5.0.1
celery==5.3.4

# Authentication & Security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation Dependencies
email-validator==2.1.0
pytz==2023.3  # For timezone validation
python-crontab==3.0.0  # For cron expression validation

# AI/ML Dependencies  
openai==1.3.5
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
nltk==3.8.1
textblob==0.17.1

# API and Integration
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email and Notifications
emails==0.6
sendgrid==6.11.0
twilio==8.10.0
slack-sdk==3.26.1

# Task Scheduling
apscheduler==3.10.4
croniter==2.0.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
python-decouple==3.8
cryptography>=41.0.0
bcrypt==4.1.1

# Monitoring and Logging
loguru==0.7.2

# Database Migration
alembic==1.12.1

# Environment and Configuration
python-dotenv==1.0.0
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Database
asyncpg==0.29.0
databases[postgresql]==0.8.0